backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
openai = "^1.95.1"
python-dotenv = "^1.1.1"
pydantic = "^2.11.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
